% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{data_to_adj}
\alias{data_to_adj}
\title{Function short description}
\usage{
data_to_adj(data, adj_fun = NULL)
}
\arguments{
\item{data}{is the data matrix.}

\item{adj_fun}{allows the users to specify their own function used to
construct the adjacency matrix. The default function is Pearson correlation \code{cor(data, use = "p")}.}
}
\value{
The adjacency matrix.
}
\description{
Function short description
}
\keyword{adjacency}
