% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utility.R
\name{eval_res}
\alias{eval_res}
\title{Function short description}
\usage{
eval_res(memberships, eval_fun = NULL)
}
\arguments{
\item{memberships}{is a list of membership vectors}

\item{eval_fun}{allows the users to specify their own function used to
evaluate clustering results. The default function is adjusted Rand index \code{mclust::adjustedRandIndex()}.}
}
\value{
A matrix containing the pairwise evaluation results between membership vectors.
}
\description{
Function short description
}
\references{
\insertCite{hubert1985comparing}{SIGNET}.
}
\keyword{evaluation}
