% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdjacencyMatrix.R
\name{getAdjacencyMatrix}
\alias{getAdjacencyMatrix}
\title{Build the adjacency matrix of a graph}
\usage{
getAdjacencyMatrix(pathway, directed = FALSE, selfLoops = FALSE)
}
\arguments{
\item{pathway}{A \verb{graphNEL} object.}

\item{directed}{If \verb{TRUE}, edges direction will be considered.}
}
\value{
A n x n matrix, n is the number of nodes in the graph.
}
\description{
This function generates the adjacency matrix from a \verb{graphNEL} object.
}
\examples{
require(signet)
data(keggPathways)
#adj<-getAdjacencyMatrix(keggPathways[[1]])
}
\keyword{adjacency}
\keyword{graph}
\keyword{matrix,}

