% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_human}
\alias{kegg_human}
\title{Pathway examples.
See \code{browseVignettes("signet")} to see how to prepare the pathways for
running simulated annealing.}
\format{An object of class \code{list} of length 3.}
\usage{
kegg_human
}
\description{
Pathway examples.
See \code{browseVignettes("signet")} to see how to prepare the pathways for
running simulated annealing.
}
\keyword{datasets}
