% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/keggPathways.R
\docType{data}
\name{keggPathways}
\alias{keggPathways}
\alias{kegg}
\title{KEGG pathways}
\format{A list with 244 graphNEL objects}
\source{
graphite
}
\usage{
kegg
}
\description{
A dataset containing 244 KEGG Pathways in graphNEL format.
}
\keyword{datasets}
