% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotAlgorithmRun.R
\name{plotAlgorithmRun}
\alias{plotAlgorithmRun}
\title{Generate an animation representing the subnetwork research}
\usage{
plotAlgorithmRun(pathway, scores, nullDist, iterations = 1000,
  temperature = 0.99, kmin = 5, directed = FALSE, verbose = TRUE,
  animPlot = 100)
}
\arguments{
\item{pathway}{A graph object.}

\item{scores}{a data frame with gene list and associated scores}

\item{nullDist}{A data frame with three columns : k, mu, sigma. Can be obtained thanks to nulldistribution() function}

\item{iterations}{The number of iterations for simulated annealings}

\item{temperature}{The temperature function parameter.}

\item{kmin}{The minimal size of a subnetwork.}

\item{directed}{If TRUE, considers the edges direction, i.e. cannot go back.}

\item{verbose}{If TRUE, displays text in the R console.}

\item{animPlot}{Number of iterations represented in the animation.}
}
\value{
A HTML file and three folder in the animation subfolder in your working directory.
}
\description{
This function will create an HTML file to visualize a run.
}
\examples{
require(signet)
data(keggPathways)
data(zScores)

#plotAlgorithmRun(keggPathways[[1]],zScores,animPlot=10)
}
\keyword{algorithm}
\keyword{annealing,}
\keyword{heuristics,}
\keyword{search}
\keyword{simulated}
\keyword{subnetwork,}

