% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotCytoscape.R
\name{plotCytoscape}
\alias{plotCytoscape}
\title{Build adjacency matrix of a graph}
\usage{
plotCytoscape(pathway, outAlgo, layout = "jgraph-spring")
}
\arguments{
\item{pathway}{A \verb{graphNEL} object.}

\item{outAlgo}{Simulated annealing output.}

\item{layout}{Layout in Cytoscape.}
}
\value{
A plot in Cytoscape.
}
\description{
This function generates the adjacency matrix from a \verb{graphNEL} object.
}
\examples{
require(signet)
data(keggPathways)
}
\keyword{Cytoscape}
\keyword{plot,}

