% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{scores}
\alias{scores}
\title{Example of gene selection scores.}
\format{A data frame with 17918 rows and 2 variables:
\describe{
  \item{gene}{Gene identifier (Entrez gene ID)}
  \item{score}{Gene selection score}
  ...
}}
\source{
Daub et al., 2013, MBE.
}
\usage{
scores
}
\description{
A dataset of gene selection scores.
}
\keyword{datasets}
