% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSubnet.R
\name{searchSubnet}
\alias{searchSubnet}
\title{Search for a high scoring subnetwork}
\usage{
searchSubnet(pathway, scores, background, iterations = 5000, verbose = TRUE)
}
\arguments{
\item{pathway}{A gene network, or a list of networks,
in the \verb{graphNEL} format.}

\item{scores}{A data frame with two columns: gene identifiers list
and associated scores.}

\item{background}{A data frame with three columns : k, mu, sigma.}

\item{iterations}{Number of iterations ofr simulated annealing.}

\item{verbose}{If \verb{TRUE}, displays text in the R console.}
}
\value{
A list containing a table with genes, their state, their score;
the subnetwortk score and size and the p-value
}
\description{
A simulated annealing algorithm to find the highest scoring
subnetwork in a graph.
}
\examples{
\dontrun{
data(daub2013)

example <- searchSubnet(kegg_human[1:5], scores, iterations = 1000)
}
}
\keyword{algorithm}
\keyword{annealing,}
\keyword{heuristics,}
\keyword{search}
\keyword{simulated}
\keyword{subnetwork,}
