% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-methods.R
\docType{class}
\name{Signet-class}
\alias{Signet-class}
\alias{.Signet}
\alias{show,Signet-method}
\alias{summary,Signet-method}
\alias{plot,Signet,missing-method}
\alias{initialize,Signet-method}
\title{An S4 class to represent a pathway and the results of the associated
simulated annealing run.}
\usage{
\S4method{show}{Signet}(object)

\S4method{summary}{Signet}(object)

\S4method{plot}{Signet,missing}(x, y, ...)

\S4method{initialize}{Signet}(.Object, pathway, scores, iterations)
}
\arguments{
\item{object}{A signet object.}

\item{x}{A signet object.}

\item{y}{Omitted when plotting a Signet object.}

\item{...}{Other graphical parameters.}

\item{.Object}{Object to initialize.}

\item{pathway}{Biological pathway (graphNEL object).}

\item{scores}{Gene scores list.}

\item{iterations}{Number of simulated annealing iterations.}
}
\value{
A plot of the simulated annealing run.

A signet object.
}
\description{
An S4 class to represent a pathway and the results of the associated
simulated annealing run.
}
\section{Methods (by generic)}{
\itemize{
\item \code{show}: Print the summary a Signet object

\item \code{summary}: Print the summary of a Signet object

\item \code{plot}: Plot a Signet object

\item \code{initialize}: Initialize a Signet object
}}

\section{Slots}{

\describe{
\item{\code{connected_comp}}{A graphNEL object (biological pathway)}

\item{\code{network}}{A data frame (gene IDs and scores)}

\item{\code{SA}}{A data frame (information on the simulated annealing run)}

\item{\code{subnet_score}}{A numeric value (subnetwork score)}

\item{\code{aggregate_score}}{A numeric value (aggregate subnetwork score)}

\item{\code{mean_score}}{A numeric value (average gene score in the pathway)}

\item{\code{subnet_size}}{An integer value (subnetwork size)}

\item{\code{subnet_genes}}{A factor (subnetwork genes)}

\item{\code{p.value}}{A numeric value (empirical p-value)}
}}

