% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/signet.R
\name{signet}
\alias{signet}
\title{Signed Network Spectral Clustering}
\usage{
signet(
  adj_mat,
  K = NULL,
  Kmax = NULL,
  Kmin = 2,
  method = c("eigen-gap", "sil-max", "sil-gap"),
  dist_fun = NULL
)
}
\arguments{
\item{adj_mat}{is an adjacency matrix.}

\item{K}{is the pre-specified number of clusters.}

\item{Kmax}{is the maximum candidate value for the number of clusters if \code{K}
is unknown.}

\item{Kmin}{is the minimum candidate value for the number of clusters if \code{K}
is unknown. The default value is \code{2}.}

\item{method}{allows the users to select the method for selecting the number of
clusters among \code{eigen-gap} (eigen-gap heuristic), \code{sil-max} (maximum
silhouette score) and \code{sil-gap} (silhouette-based gap criterion).}

\item{dist_fun}{allows the users to specify their own distance function used to
calculate the silhouette score. The default function is \code{1 - adj_mat}.}
}
\value{
A list containing:
\describe{
\item{method}{The method used to select the number of clusters.}
\item{K_sel}{The selected number of clusters.}
\item{cluster_res}{Cluster assignment under \code{K_sel}.}
}
}
\description{
Function short description
}
\examples{
library(SIGNET)

## random seed for reproducibility
set.seed(314159)

## adjacency matrix
p <- 50
mydat <- matrix(rnorm(1000 * p), ncol = p)
myadj_mat <- cor(mydat)

## implementation of signet
myres <- signet(adj_mat = myadj_mat, Kmax = 10, method = "sil-gap")
}
\references{
\insertCite{kunegis2010spectral,teng2026signet}{SIGNET}.
}
\keyword{clustering,}
\keyword{network,}
\keyword{silhouette}
