% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes-methods.R
\docType{class}
\name{SignetList-class}
\alias{SignetList-class}
\alias{.SignetList}
\alias{initialize,SignetList-method}
\alias{[[,SignetList-method}
\alias{[[}
\alias{[,SignetList,ANY,missing-method}
\alias{summary,SignetList-method}
\title{An S4 class to represent a list of "Signet" objects.}
\usage{
\S4method{initialize}{SignetList}(.Object, list)

\S4method{[[}{SignetList}(x, i)

\S4method{[}{SignetList,ANY,missing}(x, i, j, ..., drop = TRUE)

\S4method{summary}{SignetList}(object, ...)
}
\arguments{
\item{.Object}{A SignetList object.}

\item{list}{A list of Signet objects.}

\item{x}{A SignetList object.}

\item{i}{Index specifying elements to extract or replace.}

\item{j}{Unused for SignetList objects.}

\item{...}{Unused for SignetList objects.}

\item{drop}{Unused for SignetList objects.}

\item{object}{A SignetList object.}
}
\value{
Results of the simulated annealing run for a lisst of pathways.

A SignetList object.

A data frame containing
summary statistics for each element (network and subnetwork sizes,
subnetwork score, p-value, significant genes list)
}
\description{
An S4 class to represent a list of "Signet" objects.
}
\section{Methods (by generic)}{
\itemize{
\item \code{initialize}: Initialize a SignetList

\item \code{[[}: Access the ith element (signet object) of the
SignetList

\item \code{[}: Access the ith element (signet object) of the
SignetList

\item \code{summary}: Summarize the SignetList.
}}

\section{Slots}{

\describe{
\item{\code{results}}{A list of Signet objects.}
}}

