% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temperatureFunction.R
\name{temperatureFunction}
\alias{temperatureFunction}
\title{Generate the temperature function}
\usage{
temperatureFunction(iterations, param, burnin)
}
\arguments{
\item{iterations}{Number of iterations.}

\item{param}{Geometric parameter}

\item{burnin}{Burnin}
}
\description{
A function to generate the temperature function used in simulated annealing.
After a burnin period, the temperature decreases geometrically.
}
\examples{
t<-temperatureFunction(iterations=1500,param=0.995,burnin=100)
plot(t)
}
\keyword{annealing}
\keyword{simulated}
\keyword{subnetwork,}

