% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zScores.R
\name{zScores}
\alias{zScores}
\title{z-values for each gene.}
\format{A data frame with 17000 rows and 2 variables}
\source{
Daub et al. (2013)
}
\description{
A dataset containing the differentiation index over 53 human populations
of almost 18,000 genes. The variables are as follows:
}
\details{
\itemize{
  \item gene. Entrez gene ID
  \item score. The scaled Fst
  ...
}
}

