% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backgroundDist.R
\name{backgroundDist}
\alias{backgroundDist}
\title{Background distribution}
\usage{
backgroundDist(pathwaysList, scores, kmin = 1, kmax = 100,
  iterations = 1000)
}
\arguments{
\item{pathwaysList}{A graphNEL object, or a list of graphNEL objects.}

\item{scores}{A data frame in which the first column if the gene list
and the second column contains the gene scores.}

\item{kmin}{Minimal value of k for which a background distribution is
generated. Default value is 1.}

\item{kmax}{Maximal value of k for which a background distribution is
generated. Default value is the size of the biggest graph in the
provided list.}

\item{iterations}{Number of iterations to make the background distribution.}
}
\value{
A data frame containing, for each possible subnetwork size k,
the mean and standard deviation of randomly sampled subnetwork scores.
}
\description{
Generate the background distribution necessary to compute subnetwork scores
during a simulated annealing run.
}
\examples{
data(daub2013) #pathways and gene scores from Daub et al. (2013).
\dontrun{
example <- backgroundDist(kegg_human, scores)
}
}
\keyword{annealing,}
\keyword{background}
\keyword{distribution}
\keyword{simulated}
