% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/backgroundDist.R
\name{backgroundDist}
\alias{backgroundDist}
\title{Background distribution}
\usage{
backgroundDist(pathwaysList, scores, kmin = 1, kmax, iterations = 1000,
  distribution = FALSE)
}
\arguments{
\item{pathwaysList}{A graphNEL object, or a list of graphNEL objects.}

\item{scores}{A data frame in which the first column if the gene list
and the second column contains the gene scores.}

\item{kmin}{Minimal value of k for which a background distribution is
generated. Default value is 1.}

\item{kmax}{Maximal value of k for which a background distribution is
generated. Default value is the size of the biggest graph in the
provided list.}

\item{iterations}{Number of iterations to make the bachground distribution.}

\item{distribution}{If TRUE, returns the whole simulated data.
Else, returns the mean and standard deviation of the distribution for each k.}
}
\description{
Generate the background distribution necessary to compute subnetwork scores
during a simulated annealing run.
}
\examples{
require(signet)
#get data
#backgroundDist()
}
\keyword{annealing}
\keyword{simulated}
\keyword{subnetwork,}

