% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kegg_human}
\alias{kegg_human}
\title{Pathways ecample.}
\format{A list of graphNEL objects}
\usage{
kegg_human
}
\description{
Human KEGG pathways in the graphNEL format (entrez gene identifiers).
}
\keyword{datasets}
