% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/keggPathways.R
\name{keggPathways}
\alias{keggPathways}
\title{KEGG pathways}
\format{A list with 244 graphNEL objects}
\source{
graphite
}
\description{
A dataset containing 244 KEGG Pathways in graphNEL format.
}

