% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullDist.R
\name{nullDist}
\alias{nullDist}
\title{Null distribution}
\usage{
nullDist(pathways, scores, n = 1000, background)
}
\arguments{
\item{pathways}{A list of graphNEL objects.}

\item{scores}{A data frame in which the first column is the gene list
and the second column contains the gene scores.}

\item{n}{Number of null high-scores to compute.}

\item{background}{Background distribution.}
}
\description{
Generate the high-scores null distribution to compute empirical p-values
for each pathway tested.
}
\examples{
\dontrun{
data(daub13)

nullDistrib <- nullDist(kegg_human, scores, n = 1000)
}
}
\keyword{annealing,}
\keyword{distribution}
\keyword{null}
\keyword{simulated}
