% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nullDist.R
\name{nullDist}
\alias{nullDist}
\title{Null distribution}
\usage{
nullDist(pathwaysList, scores, iterations = 1000, bkgd)
}
\arguments{
\item{pathwaysList}{A list of graphNEL objects.}

\item{scores}{A data frame in which the first column if the gene list
and the second column contains the gene scores.}

\item{iterations}{Number of null high-scores to compute.}
}
\description{
Generate the high-scores null distribution to compute empirical p-values
for each pathway tested.
}
\keyword{annealing,}
\keyword{distribution}
\keyword{null}
\keyword{simulated}
