% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/nullDistribution.R
\name{nullDistribution}
\alias{nullDistribution}
\title{Null distribution}
\usage{
nullDistribution(pathwaysList, scores, kmin, kmax, iterations)
}
\arguments{
\item{pathwaysList}{A graph object, or a liste of graphNEL objects.}

\item{scores}{A data frame with gene list and associated scores}

\item{kmin}{kmin}

\item{kmax}{kmax}

\item{iterations}{Number of iterations.}
}
\description{
Generate the temperature function used in simulated annealing.
}
\examples{
require(signet)
#get data
#nullDistribution()
}
\keyword{annealing}
\keyword{simulated}
\keyword{subnetwork,}

