% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotCytoscape.R
\name{plotCytoscape}
\alias{plotCytoscape}
\title{Plot the results in Cytoscape}
\usage{
plotCytoscape(pathway, outAlgo, layout = "jgraph-spring")
}
\arguments{
\item{pathway}{A \verb{graphNEL} object.}

\item{outAlgo}{Simulated annealing output.}

\item{layout}{Layout in Cytoscape.}
}
\value{
A plot in Cytoscape.
}
\description{
This function plots the output of the algorithm in Cytoscape.
CytoscapeRPC plugin must be installed and launched in Cytoscape (< 3.0)
before running this function.
}
\examples{
require(signet)
data(keggPathways)
}
\keyword{Cytoscape}
\keyword{plot,}

