% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/plotSubnet.R
\name{plotSubnet}
\alias{plotSubnet}
\title{plotSubnet}
\usage{
plotSubnet(path, liste)
}
\arguments{
\item{pathway}{A graph object.}

\item{directed}{Set to TRUE if you want the graph to be directed.}
}
\description{
This function allows you to get the adjacency matrix from a graph object.
}
\examples{
#get data
plotSubnet()
}
\keyword{adjacency}
\keyword{matrix}

