% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSubnet.R
\name{searchSubnet}
\alias{searchSubnet}
\title{Search for a high scoring subnetwork}
\usage{
searchSubnet(pathway, scores, iterations = 1000, background)
}
\arguments{
\item{pathway}{A gene network, or a list of gene networks,
in the \verb{graphNEL} format.}

\item{scores}{A data frame with two columns: gene identifiers list (IDs have
to be the same as for the pathways, e.g. Entrez) and associated scores.}

\item{iterations}{Number of iterations.}

\item{background}{For development purposes.}
}
\value{
A \verb{signet} object or a list of \verb{signet} objects. Each
\verb{signet} object consists in a table with gene IDs, their state,
their score; the subnetwork score and size and the p-value.
}
\description{
A simulated annealing algorithm to find the highest scoring
subnetwork within a graph.
}
\examples{

# Get KEGG pathways from the package graphite:
library(graphite)
kegg <- pathways("hsapiens", "kegg")
kegg_human <- lapply(kegg[1:5], pathwayGraph)

data(daub13) # load the gene scores from Daub et al. (2013)

#run the search in all the pathways with 2500 iterations (default)
example <- searchSubnet(kegg_human, scores)
summary(example)
}
\keyword{algorithm}
\keyword{annealing,}
\keyword{heuristics,}
\keyword{search}
\keyword{simulated}
\keyword{subnetwork,}
