% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/searchSubnet.R
\name{searchSubnet}
\alias{searchSubnet}
\title{Search for a high scoring subnetwork}
\usage{
searchSubnet(pathway, scores, nullDist, iterations = 2000,
  temperature = 0.995, kmin = 3, directed = FALSE, verbose = TRUE,
  animPlot = 0)
}
\arguments{
\item{pathway}{A graph object.}

\item{scores}{a data frame with gene list and associated scores}

\item{nullDist}{A data frame with three columns : k, mu, sigma.
Can be obtained thanks to nulldistribution() function}

\item{iterations}{Number of iterations.}

\item{temperature}{The temperature function parameter.}

\item{kmin}{The minimal size of a subnetwork.}

\item{directed}{If TRUE, considers the edges direction, i.e. cannot go back.}

\item{verbose}{If TRUE, displays text in the R console.}

\item{animPlot}{If TRUE, generates a HTML file}
}
\value{
A list containing a table with genes, their state, their score;
the subnetwortk score and size and the p-value
}
\description{
A simulated annealing algorithm is used to research a high scoring
subnetwork.Implements also a test id a null distribution is provided.
}
\examples{
require(signet)
data(keggPathways)
data(zScores)

example1 <- searchSubnet(keggPathways[[1]],zScores)

#data() #a previously generated null distribution
#example2 <- searchSubnet(keggPathways[[1]],zScores,null)
}
\keyword{algorithm}
\keyword{annealing,}
\keyword{heuristics,}
\keyword{search}
\keyword{simulated}
\keyword{subnetwork,}

