% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSubnet.R
\name{searchSubnet}
\alias{searchSubnet}
\title{Search for a high scoring subnetwork}
\usage{
searchSubnet(pathway, scores, nullDist, replicates = 1, iterations = 2000,
  temperature = 0.995, subnetScore = "sum", kmin = 4, directed = FALSE,
  verbose = TRUE, burnin = 100, animPlot = 0, diagnostic = FALSE,
  dev.diag = TRUE)
}
\arguments{
\item{pathway}{A gene network, or a list of networks,
in the \verb{graphNEL} format.}

\item{scores}{A data frame with two columns: gene identifiers list
and associated scores.}

\item{nullDist}{A data frame with three columns : k, mu, sigma.
Can be obtained thanks to the \verb{nulldistribution()} function.}

\item{replicates}{Number of replicates per network.}

\item{iterations}{Number of iterations.}

\item{temperature}{The temperature function parameter.}

\item{kmin}{The minimal size of a subnetwork.}

\item{directed}{If \verb{TRUE}, considers the edges direction, i.e. cannot go back.}

\item{verbose}{If \verb{TRUE}, displays text in the R console.}

\item{burnin}{Number of iterations before the temperature begins to decrease.}

\item{animPlot}{For development purposes.}

\item{diagnostic}{If TRUE, plots the evolution of different stats along the run.}
}
\value{
A list containing a table with genes, their state, their score;
the subnetwortk score and size and the p-value
}
\description{
A simulated annealing algorithm is used to research a high scoring
subnetwork.Implements also a test id a null distribution is provided.
}
\examples{
require(signet)
data(keggPathways)
data(zScores)

example1 <- searchSubnet(keggPathways[[1]],zScores)

#data() #a previously generated null distribution
#example2 <- searchSubnet(keggPathways[[1]],zScores,null)
}
\keyword{algorithm}
\keyword{annealing,}
\keyword{heuristics,}
\keyword{search}
\keyword{simulated}
\keyword{subnetwork,}

