% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchSubnet.R
\name{searchSubnet}
\alias{searchSubnet}
\title{Search for a high scoring subnetwork}
\usage{
searchSubnet(pathway, scores, background, replicates = 1, iterations = 5000,
  subnetScore = "ideker", kmin = 2, verbose = TRUE)
}
\arguments{
\item{pathway}{A gene network, or a list of networks,
in the \verb{graphNEL} format.}

\item{scores}{A data frame with two columns: gene identifiers list
and associated scores.}

\item{background}{A data frame with three columns : k, mu, sigma.}

\item{replicates}{Number of replicates per network.}

\item{iterations}{Number of iterations.}

\item{kmin}{The minimal size of a subnetwork.}

\item{verbose}{If \verb{TRUE}, displays text in the R console.}
}
\value{
A list containing a table with genes, their state, their score;
the subnetwortk score and size and the p-value
}
\description{
A simulated annealing algorithm to find the highest scoring
subnetwork in a graph.
}
\examples{
require(signet)
data(keggPathways)
data(zScores)

example1 <- searchSubnet(keggPathways[[1]], zScores)
}
\keyword{algorithm}
\keyword{annealing,}
\keyword{heuristics,}
\keyword{search}
\keyword{simulated}
\keyword{subnetwork,}
