\name{signet}
% ----- \Rdversion{1.1}
\alias{signet}

\docType{package}

\title{Selection Inference in Gene Networks}

\description{
  Implements a method to perform inferences of selection in biological pathways. In particular it implements
a simulated annealing approach that searchs for a high scoring subnetwork in a pathway and test for its significance.
}


\examples{
## The following example illustrates the workflow.

\dontrun{
 ## (1) The user must provide...

}

}
\keyword{ package }
