\name{signet}
% ----- \Rdversion{1.1}
\alias{signet}

\docType{package}

\title{Selection Inference in Gene NETworks}

\description{
    Implements a simulated annealing approach to search for
    high scoring subnetworks of genes within biological pathways and to test
    for their significance.
}

\examples{
    ## A complete workflow is described in signet vignette:
    vignette("signet")
}
\keyword{ package }
