% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temperatureFunction.R
\name{temperatureFunction}
\alias{temperatureFunction}
\title{Temperature function}
\usage{
temperatureFunction(iterations, param, burnin)
}
\arguments{
\item{iterations}{Numbver of iterations.}

\item{param}{a data frame with gene list and associated scores}

\item{burnin}{Burnin}
}
\description{
Generate the temperature function used in simulated annealing.
Follows a geometric law, after a burnin.
}
\examples{
t<-temperatureFunction(iterations=1500,param=0.995,burnin=100)
plot(t)
}
\keyword{annealing}
\keyword{simulated}
\keyword{subnetwork,}

