% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/temperatureFunction.R
\name{temperatureFunction}
\alias{temperatureFunction}
\title{Temperature function}
\usage{
temperatureFunction(iterations, param, burnin)
}
\arguments{
\item{iterations}{A graph object.}

\item{param}{a data frame with gene list and associated scores}

\item{burnin}{}
}
\description{
Generate the temperature function used in simulated annealing.
}
\examples{
#get data
temperatureFunction()
}
\keyword{annealing}
\keyword{simulated}
\keyword{subnetwork,}

