% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/temperatureFunction.R
\name{temperatureFunction}
\alias{temperatureFunction}
\title{Generate the temperature function}
\usage{
temperatureFunction(iterations, threshold = 0.02)
}
\arguments{
\item{iterations}{Number of iterations.}

\item{threshold}{Temperature desired at the last iteration.}
}
\description{
A function to generate the temperature function used in simulated annealing.
The temperature starts from 1 and decreases geometrically for a given number
of iterations.
}
\examples{
t <- temperatureFunction(iterations = 5000)
plot(t, ylab = "Temperature", xlab = "Iterations")
}
\keyword{annealing,}
\keyword{simulated}
\keyword{subnetwork,}
\keyword{temperature}

