% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/testSubnet.R
\name{testSubnet}
\alias{testSubnet}
\title{Test the significance of a subnetwork}
\usage{
testSubnet(outputSA, pathwaysList, scores)
}
\arguments{
\item{outputSA}{Output of the algorithm.}

\item{pathwaysList}{List of pathways to generate null distribution.}
}
\description{
A function to generate the temperature function used in simulated annealing.
After a burnin period, the temperature decreases geometrically.
}
\examples{
#t<-temperatureFunction(iterations=1500,param=0.995,burnin=100)
#plot(t)
}
\keyword{annealing}
\keyword{simulated}
\keyword{subnetwork,}

