% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/testSubnet.R
\name{testSubnet}
\alias{testSubnet}
\title{Test the significance of a subnetwork}
\usage{
testSubnet(sigObj, nullD)
}
\description{
Test the significance of a subnetwork
}
\keyword{annealing}
\keyword{distribution,}
\keyword{null}
\keyword{simulated}
